backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
numpy==1.26.4
librosa==0.10.2
pydub==0.25.1
python-multipart==0.0.9
pymongo==4.7.2
bcrypt==4.1.3
python-dotenv==1.0.1
email-validator==2.1.1
PyJWT==2.8.0
